raw_map
Struct RawMap

pub struct RawMap {

    pub name: MapName,
    pub streets: StreetNetwork,
    pub buildings: BTreeMap<OsmID, RawBuilding>,
    pub areas: Vec<RawArea>,
    pub parking_lots: Vec<RawParkingLot>,
    pub parking_aisles: Vec<(WayID, Vec<Pt2D>)>,
    pub transit_routes: Vec<RawTransitRoute>,
    pub census_zones: Vec<(Polygon, CensusZone)>,
    pub transit_stops: BTreeMap<String, RawTransitStop>,
    pub bus_routes_on_roads: MultiMap<WayID, String>,
    pub osm_tags: BTreeMap<WayID, Tags>,
    pub extra_road_data: BTreeMap<RoadID, ExtraRoadData>,
    pub elevation_per_intersection: BTreeMap<IntersectionID, Distance>,
    pub extra_pois: Vec<ExtraPOI>,

}

Fields
name: MapNamestreets: StreetNetworkbuildings: BTreeMap<OsmID, RawBuilding>areas: Vec<RawArea>parking_lots: Vec<RawParkingLot>parking_aisles: Vec<(WayID, Vec<Pt2D>)>transit_routes: Vec<RawTransitRoute>census_zones: Vec<(Polygon, CensusZone)>transit_stops: BTreeMap<String, RawTransitStop>bus_routes_on_roads: MultiMap<WayID, String>

Per road, what bus routes run along it?

This is scraped from OSM relations for every map, unlike the more detailed transit_routes above, which come from GTFS only for a few maps. This is used only to identify roads part of bus routes. It’s best-effort and not robust to edits or transformations.

osm_tags: BTreeMap<WayID, Tags>extra_road_data: BTreeMap<RoadID, ExtraRoadData>elevation_per_intersection: BTreeMap<IntersectionID, Distance>extra_pois: Vec<ExtraPOI>
Implementations

impl RawMap

pub fn blank(name: MapName) -> RawMap

pub fn save(&self)

pub fn get_city_name(&self) -> &CityName

pub fn road_to_osm_tags(&self, id: RoadID) -> Option<&Tags>