Crate raw_map

The convert_osm crate produces a RawMap from OSM and other data. Storing this intermediate structure is useful to iterate quickly on parts of the map importing pipeline without having to constantly read .osm files, and to visualize the intermediate state with map_editor.

Modules

types 🔒

Structs

Amenity

A business located inside a building.

ExtraPOI

Extra point-of-interest

ExtraRoadData

Extra data associated with one Road

RawArea

RawBuilding

RawMap

RawParkingLot

RawTransitRoute

RawTransitStop

Enums

AmenityType

Businesses are categorized into one of these types.

AreaType

CrossingType

Classifies pedestrian and cyclist crossings. Note lots of detail is missing.

ExtraPOIType

RawTransitType