Crate raw_map

The convert_osm crate produces a RawMap from OSM and other data. Storing this intermediate structure is useful to iterate quickly on parts of the map importing pipeline without having to constantly read .osm files, and to visualize the intermediate state with map_editor.

Modules

geometry

OSM describes roads as center-lines that intersect. Turn these into road and intersection polygons roughly by

initial

Naming is confusing, but RawMap -> InitialMap -> Map. InitialMap is separate pretty much just for the step of producing https://a-b-street.github.io/docs/tech/map/importing/geometry.html.

lane_specs

osm

Useful utilities for working with OpenStreetMap.

transform

types

Structs

Amenity

A business located inside a building.

LaneSpec

MapConfig

NamePerLanguage

None corresponds to the native name

OriginalRoad

A way to refer to roads across many maps and over time. Also trivial to relate with OSM to find upstream problems.

RawArea

RawBuilding

RawIntersection

RawMap

RawParkingLot

RawRoad

RawTransitRoute

RawTransitStop

TurnRestriction

Enums

AmenityType

Businesses are categorized into one of these types.

AreaType

BufferType

Direction

DrivingSide

IntersectionType

LaneType

RawTransitType

RestrictionType

Constants

NORMAL_LANE_THICKNESS

SIDEWALK_THICKNESS

Functions

get_lane_specs_ltr

intersection_polygon

Also returns a list of labeled polygons for debugging.