Module raw_map::geometry

OSM describes roads as center-lines that intersect. Turn these into road and intersection polygons roughly by

treating the road as a PolyLine with a width, so that it has a left and right edge

finding the places where the edges of different roads intersect

“Trimming back” the center lines to avoid the overlap

Producing a polygon for the intersection itsef

I wrote a novella about this: https://a-b-street.github.io/docs/tech/map/geometry/index.html

Modules

algorithm

geojson

Structs

InputRoad

Results

Functions

intersection_polygon

Also returns a list of labeled polygons for debugging.

intersection_polygon_v2

osm2polygon

Process the file produced by save_osm2polygon_input, then write the output as GeoJSON.