Struct raw_map::RawMap

pub struct RawMap {
    pub name: MapName,
    pub streets: StreetNetwork,
    pub buildings: BTreeMap<OsmID, RawBuilding>,
    pub areas: Vec<RawArea>,
    pub parking_lots: Vec<RawParkingLot>,
    pub parking_aisles: Vec<(WayID, Vec<Pt2D>)>,
    pub transit_routes: Vec<RawTransitRoute>,
    pub transit_stops: BTreeMap<String, RawTransitStop>,
    pub bus_routes_on_roads: MultiMap<WayID, String>,
}

Fields
name: MapNamestreets: StreetNetworkbuildings: BTreeMap<OsmID, RawBuilding>areas: Vec<RawArea>parking_lots: Vec<RawParkingLot>parking_aisles: Vec<(WayID, Vec<Pt2D>)>transit_routes: Vec<RawTransitRoute>transit_stops: BTreeMap<String, RawTransitStop>bus_routes_on_roads: MultiMap<WayID, String>

Per road, what bus routes run along it?

This is scraped from OSM relations for every map, unlike the more detailed transit_routes above, which come from GTFS only for a few maps. This is used only to identify roads part of bus routes. It’s best-effort and not robust to edits or transformations.

Implementations

impl RawMap

pub fn blank(name: MapName) -> RawMap

pub fn new_osm_way_id(&self, start: i64) -> WayID

pub fn new_osm_node_id(&self, start: i64) -> NodeID

pub fn save(&self)

pub fn get_city_name(&self) -> &CityName